requirements.txt
wheel
cmake
pyngrok
fastapi 
uvicorn 

# google translate api
googletrans==4.0.0-rc1
langdetect

# spacy for comparing text
spacy
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
<<<<<<< HEAD
epsilon==0.8.0
essential-generators==1.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
facexlib==0.3.0
fairscale==0.4.13
faiss-cpu==1.7.3
fastapi==0.109.2
filelock==3.13.1
filterpy==1.4.5
fire==0.5.0
Flask==3.0.2
flask-babel==4.0.0
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
ftfy==6.1.3
future==0.18.3
gast==0.5.4
glob2==0.7
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
googletrans==4.0.0rc1
graphql-core==3.2.3
grpcio==1.60.1
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.20.3
humanfriendly==10.0
humanize==4.9.0
Hypercorn==0.16.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
image-similarity-measures==0.3.6
imageio==2.33.1
imgaug==0.4.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
imutils==0.5.4
incremental==22.10.0
inference-sdk==0.9.14
inflate64==1.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonlines==4.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kaleido==0.2.1
keras==2.15.0
Keras-Applications==1.0.8
keras-ocr==0.9.3
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
kornia==0.7.1
langcodes==3.3.0
langdetect==1.0.9
lazy_loader==0.3
libclang==16.0.6
lightning-utilities==0.10.0
lit==17.0.6
llvmlite==0.42.0
lmdb==1.4.1
lxml==5.1.0
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.5.2
mdurl==0.1.2
ml-dtypes==0.2.0
mongoengine==0.24.2
motor==3.3.2
mpmath==1.3.0
multidict==6.0.4
multivolumefile==0.2.3
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
ninja==1.11.1.1
numba==0.59.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.15.0
onnx-graphsurgeon==0.3.27
onnx2tf==1.17.5
onnxruntime==1.17.0
onnxsim==0.4.35
open-clip-torch==2.24.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.8.0.74
openpyxl==3.1.2
opt-einsum==3.3.0
paddleocr==2.7.0.3
paddlepaddle==2.6.0
pandas==1.4.3
pdf2docx==0.5.8
phasepack==1.5
pillow==10.2.0
plotly==5.18.0
pooch==1.8.1
pprintpp==0.4.0
premailer==3.10.0
preshed==3.0.9
priority==2.0.0
protobuf==3.20.3
py-cpuinfo==9.0.0
py7zr==0.20.8
pyasn1==0.5.1
pyasn1-modules==0.3.0
pybcj==1.0.2
pybind11==2.11.1
pyclipper==1.3.0.post5
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
pycryptodomex==3.20.0
pydantic==2.6.1
pydantic_core==2.16.2
pyee==8.2.2
PyMatting==1.1.12
pymongo==4.6.1
PyMuPDF==1.20.2
pyOpenSSL==24.0.0
pyparsing==2.4.7
pyppeteer==1.0.2
pyppmd==1.1.0
pyre-extensions==0.0.23
python-bidi==0.4.2
python-docx==1.1.0
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.9
pytorch-lightning==1.9.5
pytz==2023.3.post1
PyWavelets==1.5.0
PyYAML==6.0.1
pyzstd==0.15.9
rapidfuzz==3.6.1
rarfile==4.1
referencing==0.33.0
regex==2023.12.25
rembg==2.0.54
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retrying==1.3.4
rf-groundingdino==0.1.2
rf-segment-anything==1.0
rfc3986==1.5.0
rich==13.7.0
roboflow==1.1.14
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.0
safetensors==0.4.1
scikit-image==0.20.0
scikit-learn==1.3.2
scipy==1.9.1
seaborn==0.13.0
sentencepiece==0.2.0
service-identity==24.1.0
shapely==2.0.2
smart-open==6.4.0
sng4onnx==1.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sse-starlette==0.10.3
sseclient-py==1.8.0
starlette==0.36.3
strawberry-graphql==0.138.1
supervision==0.9.0
sympy==1.12
tabulate==0.9.0
taskgroup==0.0.0a4
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.15.0.post1
tensorflow-cpu==2.15.0.post1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
texttable==1.7.0
tf-estimator-nightly==2.8.0.dev2021122109
tflite-support==0.4.4
thinc==8.2.3
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==0.6.13
tokenizers==0.15.0
tomli==2.0.1
torch==2.2.1
torchaudio==0.10.0
torchdiffeq==0.2.3
torchmetrics==1.2.1
torchvision==0.17.1
tqdm==4.65.0
transformers==4.36.2
triton==2.2.0
Twisted==23.10.0
typer==0.9.0
typing==3.7.4.3
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.2
ultralytics==8.1.3
universal-analytics-python3==1.1.1
urllib3==1.26.18
uvicorn==0.27.1
validators==0.22.0
visualdl==2.5.3
voxel51-eta==0.12.4
wasabi==1.1.2
weasel==0.3.4
webcolors==1.13
websockets==10.4
Werkzeug==3.0.1
wrapt==1.14.1
wsproto==1.2.0
xformers==0.0.16
xmltodict==0.13.0
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
zope.interface==6.2
=======

# yolov8
ultralytics
torch 
torchvision 
torchaudio

# ocr
paddleocr
paddlepaddle

# face recognition 
dlib
face_recognition
python-multipart

# Search Engine
DeepImageSearch

# roboflow
roboflow
>>>>>>> 9539924130e13f32efd1725cccaf994b23fa8e1a
